vibe-monitor/requirements.txt
fastapi
uvicorn
prometheus-fastapi-instrumentator
opentelemetry-distro
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
opentelemetry-instrumentation-logging
requests